# services/content/src/api/openapi.yaml
openapi: 3.1.0
info:
  title: Content Service API
  version: 1.0.0
  description: |
    Production-grade content management service for the Suuupra platform.
    
    ## Features
    - Content CRUD operations with rich metadata
    - Large file uploads with S3 multipart and progress tracking
    - Elasticsearch-powered search and discovery
    - Content approval workflow with versioning
    - Real-time WebSocket notifications
    - Multi-tenant architecture with RBAC
    
    ## Authentication
    All endpoints require JWT authentication via the `Authorization: Bearer <token>` header.
    
    ## Rate Limiting
    API endpoints are rate limited. Check response headers for current limits.
    
    ## Versioning
    Content uses semantic versioning (major.minor.patch) with immutable history.
  contact:
    name: Suuupra Platform Team
    email: platform-team@suuupra.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:8082/api/v1
    description: Development server
  - url: https://api.suuupra.com/content/v1
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /search:
    get:
      summary: Search content
      description: Search content using Elasticsearch with faceted filtering
      operationId: searchContent
      parameters:
        - name: q
          in: query
          required: true
          description: Search query
          schema:
            type: string
            minLength: 1
            maxLength: 500
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - name: sort
          in: query
          description: Sort field
          schema:
            type: string
            enum: [relevance, created_at, updated_at, title, views]
            default: relevance
        - $ref: '#/components/parameters/Order'
        - name: filters
          in: query
          description: Search filters (JSON object)
          schema:
            type: string
            description: JSON-encoded filter object
      responses:
        '200':
          description: Search results retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    Page:
      name: page
      in: query
      description: Page number
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit:
      name: limit
      in: query
      description: Items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    Order:
      name: order
      in: query
      description: Sort order
      schema:
        type: string
        enum: [asc, desc]
        default: desc
  schemas:
...
    ContentType:
      type: string
      enum: [video, article, quiz, document]
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
        error:
          $ref: '#/components/schemas/ApiError'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
      required:
        - success
    ApiError:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
      required:
        - code
        - message
    ResponseMeta:
      type: object
      properties:
        requestId:
          type: string
        timestamp:
          type: string
          format: date-time
        pagination:
          $ref: '#/components/schemas/PaginationMeta'
      required:
        - requestId
        - timestamp
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
        hasNext:
          type: boolean
        hasPrev:
          type: boolean
      required:
        - page
        - limit
        - total
        - totalPages
        - hasNext
        - hasPrev
...
    SearchResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/SearchResult'
            meta:
              allOf:
                - $ref: '#/components/schemas/ResponseMeta'
                - type: object
                  properties:
                    aggregations:
                      type: object
                      additionalProperties: true
                    queryTimeMs:
                      type: integer
                    totalHits:
                      type: integer
    SearchResult:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        contentType:
          $ref: '#/components/schemas/ContentType'
        category:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
            path:
              type: string
        tags:
          type: array
          items:
            type: string
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
        publishedAt:
          type: string
          format: date-time
        fileInfo:
          type: object
        _score:
          type: number
        highlights:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
...
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            success:
              type: boolean
              enum: [false]
          required:
            - success
            - error
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: VALIDATION_ERROR
              message: Validation failed
              details:
                errors:
                  - field: title
                    message: must not be empty
                    value: ""
...